Catch compiler panics when testing + fix panick on empty file
The compiler panicked when being fed an empty file. This is fixed now.
Unfortunately we can't really test this as an empty file will result in
an error (no main function not found) resulting in a failing test. But if
we'd add a comment to tell the runner to expect this error, the file wouldn't
be empty any more.