[feat] Finish assembly lifetime analysis
This commit implements lifetime analysis for assembly code based on
`Linear Scan Registry Allocation on SSA Form` by Christian Wimmer and
Michael Frany.

Note that lifetime analysis panics when encountering unused arguments
or registers as this case is not really specified in the paper.
In the future this should be handled better, for now the tests are
adapted to not have unused arguments.