[feat] Introduce Stack Slots
In many cases we store variables on the stack and then load/store
them as needed. Until now RusTiny used the same type of register
for both pointers to the stack and local variables.

This commit introduces a notion of stack slots. They are basically
normal registers which contain pointers to a variable stored
on the stack. This helps keeping pointers and raw values apart
later on.